finished CRUD for publications entity

+++ PublicationsAPI/Services/PublicationsService.cs
@@ -33,9 +33,15 @@ namespace PublicationsAPI.Services {
             return (await _publicationsRepository.AddPublicationAsync(publication)).PublicationsToPublicationResponseDTO();
         }
 
-        public Task<bool> DeletePublicationAsync(string publicationUuid)
+        public async Task<bool> DeletePublicationAsync(string publicationUuid, string authorUuid)
         {
-            return _publicationsRepository.DeletePublicationAsync(publicationUuid);
+            if( !((await _publicationsRepository.GetPublicationAsync(publicationUuid)).AuthorUuid == authorUuid) )
+            {
+                return false;
+                //throw unauthenticated exception
+            }
+            else 
+                return await _publicationsRepository.DeletePublicationAsync(publicationUuid);
         }
 
         public async Task<PublicationResponseDTO> GetPublicationAsync(string publicationUuid)
@@ -44,8 +50,9 @@ namespace PublicationsAPI.Services {
         }
 
         public async Task<IEnumerable<PublicationResponseDTO>> GetPublicationsFromUserAsync(string publisherUuid)
-        {
-            return ((await _usersServices.GetPublicationsByUserUuidAsync(publisherUuid)).Select(p => p.PublicationsToPublicationResponseDTO())).ToList();
+        {   
+            var userPublications = await _usersServices.GetPublicationsByUserUuidAsync(publisherUuid);
+            return userPublications.Select(p => p.PublicationsToPublicationResponseDTO()).ToList();
         }
 
         public async Task<IEnumerable<PublicationResponseDTO>> GetPublicationsPaginatedAsync(string publisherUuid, int page, int pageSize)
@@ -56,9 +63,12 @@ namespace PublicationsAPI.Services {
             return publications.Select(publications => publications.PublicationsToPublicationResponseDTO());
         }
 
-        public async Task<PublicationResponseDTO> UpdatePublicationAsync(string publicationUuid, PublicationDTO publicationDto)
+        public async Task<PublicationResponseDTO> UpdatePublicationAsync(string publicationUuid, PublicationDTO publicationDto, string authorUuid)
         {
-            Publications? publicationToUpdate = await _publicationsRepository.GetPublicationAsync(publicationUuid);
+            Publications publicationToUpdate = await _publicationsRepository.GetPublicationAsync(publicationUuid);
+
+            if(publicationToUpdate == null)
+                throw new NullReferenceException("Any Publication with this UUID was found");
 
             publicationToUpdate.Title = publicationDto.Title;
             publicationToUpdate.Description = publicationDto.Description;
@@ -66,7 +76,12 @@ namespace PublicationsAPI.Services {
             publicationToUpdate.PublicationType = publicationDto.PublicationType;
             publicationToUpdate.UpdatedAt = DateTime.UtcNow;
 
-            return (await _publicationsRepository.UpdatePublicationAsync(publicationToUpdate)).PublicationsToPublicationResponseDTO();
+            Publications? publicationUpdated = await _publicationsRepository.UpdatePublicationAsync(publicationToUpdate);
+
+            if(publicationUpdated == null)
+                throw new ArgumentNullException("An error occured with the update requested on this publication");
+
+            return publicationUpdated.PublicationsToPublicationResponseDTO();
         }
     }
 }
